+++ tests/smoke/auto-format-tools.sh
@@ -2,4 +2,5 @@
 
 SMOKE_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
+. ${SMOKE_DIR}/base.sh "use_black=yes"
 . ${SMOKE_DIR}/base.sh "use_black=no"